Refactoring provider_iwls class
