+++ utils/compile_wheel/compile_wheel.sh
@@ -4,16 +4,11 @@ mkdir -p install/src/provider_iwls
 touch install/src/provider_iwls/__init__.py
 cp ./utils/compile_wheel/setup.py ./install
 cp ./utils/compile_wheel/README.md ./install
-cp ./pygeoapi_iwls/provider_iwls.py ./install/src/provider_iwls
-cp ./pygeoapi_iwls/s104.py ./install/src/provider_iwls
-cp ./pygeoapi_iwls/s111.py ./install/src/provider_iwls
-cp ./pygeoapi_iwls/s100.py ./install/src/provider_iwls
-cp ./pygeoapi_iwls/iwls.py ./install/src/provider_iwls
-cp ./pygeoapi_iwls/s100_util.py ./install/src/provider_iwls
-cp ./pygeoapi_iwls/process_iwls.py ./install/src/provider_iwls
+rm ./install/src/provider_iwls/*
+cp -r ./pygeoapi_iwls/* ./install/src/provider_iwls
 pip install wheel
 pip wheel ./install
-pip install --upgrade --force-reinstall provider_iwls-0.0.1-py3-none-any.whl
+pip install --upgrade --force-reinstall --debug provider_iwls-0.0.1-py3-none-any.whl
 export PYGEOAPI_CONFIG=config_dev.yml
 export PYGEOAPI_OPENAPI=config.yml
 pygeoapi openapi generate $PYGEOAPI_CONFIG > $PYGEOAPI_OPENAPI
